refactor: PaymentService -> PaymentClient로 변경
