+++ src/main/java/roomescape/client/PaymentClient.java
@@ -1,4 +1,4 @@
-package roomescape.service;
+package roomescape.client;
 
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.service.annotation.HttpExchange;
@@ -7,8 +7,8 @@ import roomescape.dto.payment.PaymentRequest;
 import roomescape.dto.payment.PaymentResponse;
 
 @HttpExchange
-public interface PaymentService {
+public interface PaymentClient {
 
     @PostExchange("/confirm")
-    PaymentResponse pay(@RequestBody PaymentRequest paymentRequest);
+    PaymentResponse requestPayment(@RequestBody PaymentRequest paymentRequest);
 }
